requirements.txt
# LinkedIn Auto Job Applier Dependencies
# Python 3.10+ required

# Environment & Configuration
python-dotenv>=1.0.0

# Browser Automation
undetected-chromedriver>=3.5.0
selenium>=4.0.0

# GUI Automation
pyautogui>=0.9.54

# Web Framework (for job history UI)
flask>=3.0.0
flask-cors>=4.0.0

# AI Integrations (optional)
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
setuptools>=69.0.0
